fix: consolidate schema and cred def fixtures
